Relocate AzureBlobStorageService and its DI to the Azure adapter
