+++ backend/src/Anela.Heblo.API/Program.cs
@@ -101,6 +101,7 @@ public partial class Program
         // Add new architecture services
         builder.Services.AddPersistenceServices(builder.Configuration, builder.Environment);
         builder.Services.AddApplicationServices(builder.Configuration, builder.Environment); // Vertical slice modules from Application layer
+        builder.Services.AddAzureBlobStorageService(builder.Environment); // Azure adapter: BlobServiceClient + IBlobStorageService binding (moved out of Application ring)
         builder.Services.AddScoped<ISmartsuppWebhookMetrics, SmartsuppWebhookMetrics>();
         builder.Services.AddXccServices(builder.Configuration); // Cross-cutting concerns (audit, telemetry, etc.)
         builder.Services.AddUsersModule(); // Users feature composition root (API-layer adapter for ICurrentUserService)
