+++ backend/src/Anela.Heblo.Application/Features/FileStorage/FileStorageModule.cs
@@ -1,8 +1,6 @@
 using System.Net;
 using Anela.Heblo.Application.Features.FileStorage.Infrastructure;
-using Anela.Heblo.Application.Features.FileStorage.Services;
 using Anela.Heblo.Domain.Features.FileStorage;
-using Azure.Storage.Blobs;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
@@ -37,25 +35,6 @@ public static class FileStorageModule
                 .ValidateOnStart();
         }
 
-        // Register Azure Blob Storage client. The factory reads the already-validated options,
-        // so ValidateOnStart() runs before any consumer resolves the BlobServiceClient.
-        services.AddSingleton<BlobServiceClient>(provider =>
-        {
-            var opts = provider.GetRequiredService<IOptions<FileStorageOptions>>().Value;
-            if (string.IsNullOrWhiteSpace(opts.BlobConnectionString))
-            {
-                // Reachable only in Development — validation blocks the empty path elsewhere.
-                // Log a warning so the storage-emulator fallback is never silent.
-                var logger = provider.GetRequiredService<ILogger<AzureBlobStorageService>>();
-                logger.LogWarning(
-                    "FileStorage:BlobConnectionString is empty in {Environment}; falling back to UseDevelopmentStorage=true.",
-                    environment.EnvironmentName);
-                return new BlobServiceClient("UseDevelopmentStorage=true");
-            }
-
-            return new BlobServiceClient(opts.BlobConnectionString);
-        });
-
         // Register named HttpClient for product export downloads.
         // PooledConnectionLifetime recycles sockets and refreshes DNS every 5 minutes,
         // preventing the stale-socket and DNS-pinning problems of a long-lived singleton HttpClient.
@@ -78,10 +57,6 @@ public static class FileStorageModule
         // Register resilience service as Singleton — it holds no request state and
         // its internal Polly pipeline is rebuilt per-call (see BuildPipeline).
         services.AddSingleton<IDownloadResilienceService, DownloadResilienceService>();
-
-        // Register blob storage service as Singleton so the _containerExists cache survives across requests.
-        // BlobServiceClient is already Singleton — no thread-safety concerns.
-        services.AddSingleton<IBlobStorageService, AzureBlobStorageService>();
 
         services.Configure<FileDownloadOptions>(configuration.GetSection("FileStorage:Download"));
